Created 'Create_Starter_Product' functions and map.
