--- Products/Cocaine.cpp
@@ -1,9 +1,0 @@
-//
-// Created by den on 6/29/25.
-//
-
-#include "Cocaine.h"
-
-
-
-Cocaine::~Cocaine() = default;
