--- Products/Granddaddy_Purple.cpp
@@ -1,9 +1,0 @@
-//
-// Created by den on 6/29/25.
-//
-
-#include "Granddaddy_Purple.h"
-
-
-
-Granddaddy_Purple::~Granddaddy_Purple() = default;
