--- Products/Green_Crack.cpp
@@ -1,9 +1,0 @@
-//
-// Created by den on 6/29/25.
-//
-
-#include "Green_Crack.h"
-
-
-
-Green_Crack::~Green_Crack() = default;
