--- Products/Methamphetamine.cpp
@@ -1,9 +1,0 @@
-//
-// Created by den on 6/29/25.
-//
-
-#include "Methamphetamine.h"
-
-
-
-Methamphetamine::~Methamphetamine() = default;
