--- Products/OG_Kush.cpp
@@ -1,9 +1,0 @@
-//
-// Created by den on 6/29/25.
-//
-
-#include "OG_Kush.h"
-
-
-
-OG_Kush::~OG_Kush() = default;
