--- Products/Products.h
@@ -1,16 +1,0 @@
-//
-// Created by den on 7/20/25.
-//
-
-#ifndef PRODUCTS_H
-#define PRODUCTS_H
-
-#include "Product.h"
-#include "Cocaine.h"
-#include "Granddaddy_Purple.h"
-#include "Green_Crack.h"
-#include "Methamphetamine.h"
-#include "OG_Kush.h"
-#include "Sour_Diesel.h"
-
-#endif //PRODUCTS_H
