+++ Products/Products_Lib.h
@@ -5,8 +5,10 @@
 #ifndef PRODUCTSLIB_H
 #define PRODUCTSLIB_H
 #include <map>
+#include <cstdint>
+#include <vector>
+#include "../Effects_Lib.h"
 class Product;
-
 
 namespace Products_Lib
 {
@@ -38,21 +40,7 @@ namespace Products_Lib
 		{starter_product::COCAINE, 150}
 	};
 
-	static Product* create_Cocaine();
-	static Product* create_Granddaddy_Purple();
-	static Product* create_Green_Crack();
-	static Product* create_Methamphetamine();
-	static Product* create_OG_Kush();
-	static Product* create_Sour_Diesel();
 
-	static const std::map<starter_product, Product* (*)()> create_Product = {
-		{starter_product::OG_KUSH, create_OG_Kush},
-		{starter_product::SOUR_DIESEL, create_Sour_Diesel},
-		{starter_product::GREEN_CRACK, create_Green_Crack},
-		{starter_product::GRANDDADDY_PURPLE, create_Granddaddy_Purple},
-		{starter_product::METHAMPHETAMINE, create_Methamphetamine},
-		{starter_product::COCAINE, create_Cocaine},
-	};
-}
+};
 
 #endif //PRODUCTSLIB_H
