--- Products/Sour_Diesel.cpp
@@ -1,9 +1,0 @@
-//
-// Created by den on 6/29/25.
-//
-
-#include "Sour_Diesel.h"
-
-
-
-Sour_Diesel::~Sour_Diesel() = default;
